+++ dictos/__init__.py
@@ -1,6 +1,6 @@
 from dictos.__version__ import __version__
 from dictos.core.expr import Expr
-from dictos.lagrangian_polynomial import lagrangian_basis, lagrangian_poly, derivative
+from dictos.poly import lagrangian_polynomial
 from dictos.taylor_expansion import taylor_series
 from dictos.calculus import finite_difference
 from dictos.poly import interpolation
